Guessing Game
This is a simple guessing game between two players. The system prompt
each player to enter integers for setting secret number and for
guessing secret number. Out of bound input is considered invalid and
will be prompted to provide another valid input. The game also keeps
track of valid guesses and provides success reminder when game ends.
The assumption for this game is that both players will provide integer
inputs only.